refactor: reorganize repository structure and update imports for agent and context repositories
