+++ src/rag_service/repositories/context/base.py
@@ -5,7 +5,6 @@ This module defines the interfaces for document context and agent repositories.
 
 from abc import ABC, abstractmethod
 
-from src.models.agent import Agent
 from src.rag_service.context import Context
 
 
@@ -55,51 +54,6 @@ class ContextRepository(ABC):
 
         Returns:
             Context: The stored context object
-        """
-
-    @abstractmethod
-    def is_reachable(self) -> bool:
-        """Check if the repository backend is accessible.
-
-        Returns:
-            bool: True if connection is healthy, False otherwise
-        """
-
-
-class AgentRepository(ABC):
-    """Abstract base class for agent storage.
-
-    This repository handles persistence of AI agent configurations.
-    """
-
-    @abstractmethod
-    def create_agent(self, agent: Agent) -> Agent:
-        """Store a new agent configuration.
-
-        Args:
-            agent (Agent): The agent object to persist
-
-        Returns:
-            Agent: The stored agent object
-        """
-
-    @abstractmethod
-    def get_agents(self) -> list[Agent]:
-        """Retrieve all stored agent configurations.
-
-        Returns:
-            list[Agent]: List of all agent configurations
-        """
-
-    @abstractmethod
-    def get_agent_by_id(self, agent_id: str) -> Agent | None:
-        """Retrieve a specific agent by ID.
-
-        Args:
-            agent_id (str): The unique identifier of the agent
-
-        Returns:
-            Agent | None: The agent if found, None otherwise
         """
 
     @abstractmethod
