+++ src/rag_service/repositories/context/mongodb_context_repository.py
@@ -5,7 +5,7 @@ from pymongo import MongoClient
 from src.config import Config
 from src.rag_service.context import Context
 from src.rag_service.embeddings import similarity_search
-from src.rag_service.repositories.base import ContextRepository
+from src.rag_service.repositories.context.base import ContextRepository
 
 
 config = Config()
