+++ tests/mocks/mock_agent_repository.py
@@ -3,7 +3,7 @@
 from copy import deepcopy
 
 from src.models.agent import Agent
-from src.rag_service.repositories.base import AgentRepository
+from src.rag_service.repositories import AgentRepository
 
 
 class MockAgentRepository(AgentRepository):
